+++ src/main/java/twilightforest/entity/EntityTFFireBeetle.java
@@ -57,7 +57,7 @@ public class EntityTFFireBeetle extends EntityMob implements IBreathAttacker
     protected void entityInit()
     {
         super.entityInit();
-        dataWatcher.addObject(17, Byte.valueOf((byte)0));
+        dataWatcher.addObject(17, (byte)0);
     }
 	
     /**
@@ -143,11 +143,11 @@ public class EntityTFFireBeetle extends EntityMob implements IBreathAttacker
     {
         if (flag)
         {
-            dataWatcher.updateObject(17, Byte.valueOf((byte)127));
+            dataWatcher.updateObject(17, ((byte)127));
         }
         else
         {
-            dataWatcher.updateObject(17, Byte.valueOf((byte)0));
+            dataWatcher.updateObject(17, (byte)0);
         }
     }
