+++ src/main/java/twilightforest/entity/EntityTFMinotaur.java
@@ -59,7 +59,7 @@ public class EntityTFMinotaur extends EntityMob implements ITFCharger {
     protected void entityInit()
     {
         super.entityInit();
-        dataWatcher.addObject(17, Byte.valueOf((byte)0));
+        dataWatcher.addObject(17, (byte)0);
     }
 	
     /**
@@ -80,11 +80,11 @@ public class EntityTFMinotaur extends EntityMob implements ITFCharger {
     {
         if (flag)
         {
-            dataWatcher.updateObject(17, Byte.valueOf((byte)127));
+            dataWatcher.updateObject(17, ((byte)127));
         }
         else
         {
-            dataWatcher.updateObject(17, Byte.valueOf((byte)0));
+            dataWatcher.updateObject(17, (byte)0);
         }
     }
