+++ src/main/java/twilightforest/entity/boss/EntityTFUrGhast.java
@@ -103,8 +103,8 @@ public class EntityTFUrGhast extends EntityTFTowerGhast implements IBossDisplayD
     protected void entityInit()
     {
         super.entityInit();
-//        this.dataWatcher.addObject(DATA_BOSSHEALTH, new Integer(this.getMaxHealth()));
-        this.dataWatcher.addObject(DATA_TANTRUM, Byte.valueOf((byte) 0));
+//        this.dataWatcher.addObject(DATA_BOSSHEALTH, (this.getMaxHealth()));
+        this.dataWatcher.addObject(DATA_TANTRUM, ((byte) 0));
     }
 
     /**
@@ -471,7 +471,7 @@ public class EntityTFUrGhast extends EntityTFTowerGhast implements IBossDisplayD
 
         if (currentAggroStatus != newAggroStatus)
         {
-        	this.dataWatcher.updateObject(16, Byte.valueOf(newAggroStatus));
+        	this.dataWatcher.updateObject(16, newAggroStatus);
         }
     }
 
@@ -796,7 +796,7 @@ public class EntityTFUrGhast extends EntityTFTowerGhast implements IBossDisplayD
      */
     public void setInTantrum(boolean par1)
     {
-    	this.dataWatcher.updateObject(DATA_TANTRUM, par1 ? Byte.valueOf((byte)-1) : Byte.valueOf((byte)0));
+    	this.dataWatcher.updateObject(DATA_TANTRUM, par1 ? ((byte)-1) : (byte)0);
     	
     	// can we just reset this each time it is called?
     	this.damageUntilNextPhase = 48;
