+++ src/main/java/twilightforest/entity/passive/EntityTFQuestRam.java
@@ -72,8 +72,8 @@ public class EntityTFQuestRam extends EntityAnimal {
 	protected void entityInit()
     {
         super.entityInit();
-        this.dataWatcher.addObject(16, Integer.valueOf(0));
-        this.dataWatcher.addObject(17, Byte.valueOf((byte)0));
+        this.dataWatcher.addObject(16, (int)0);
+        this.dataWatcher.addObject(17, (byte)0);
     }
 
 
@@ -259,7 +259,7 @@ public class EntityTFQuestRam extends EntityAnimal {
      */
     public void setColorFlags(int par1)
     {
-    	this.dataWatcher.updateObject(16, Integer.valueOf(par1));
+    	this.dataWatcher.updateObject(16, par1);
     }
 
     /**
@@ -295,7 +295,7 @@ public class EntityTFQuestRam extends EntityAnimal {
      */
     public void setRewarded(boolean par1)
     {
-    	this.dataWatcher.updateObject(17, par1 ? Byte.valueOf((byte)1) : Byte.valueOf((byte)0));
+    	this.dataWatcher.updateObject(17, par1 ? ((byte)1) : (byte)0);
     }
